requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
httpx
slowapi
cachetools
jinja2
pydantic

# Tests
pytest
pytest-asyncio
pytest-cov
httpx